Split settings into components and make labels bold
