+++ osu.Game/Screens/Multi/Screens/Match/Settings.cs
@@ -1,9 +1,7 @@
 // Copyright (c) 2007-2018 ppy Pty Ltd <contact@ppy.sh>.
 // Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
 
-using osu.Framework.Allocation;
 using osu.Framework.Configuration;
-using osu.Framework.Extensions;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Shapes;
@@ -35,7 +33,7 @@ namespace osu.Game.Screens.Multi.Screens.Match
         {
             OsuTextBox roomNameBox;
             OsuTextBox maxParticipantsBox;
-            OsuPasswordTextBox passwordBox;
+            SettingsPasswordTextBox passwordBox;
 
             Children = new Drawable[]
             {
@@ -56,6 +54,7 @@ namespace osu.Game.Screens.Multi.Screens.Match
                         {
                             Text = @"ROOM NAME",
                             Colour = Color4.White,
+                            Font = @"Exo2.0-Bold",
                         },
                         roomNameBox = new SettingsTextBox
                         {
@@ -68,8 +67,9 @@ namespace osu.Game.Screens.Multi.Screens.Match
                             Margin = new MarginPadding { Top = 100 },
                             Text = @"ROOM VISIBILITY",
                             Colour = Color4.White,
+                            Font = @"Exo2.0-Bold",
                         },
-                        roomAvailabilityTabs = new AvailabilityTabControl
+                        roomAvailabilityTabs = new RoomAvailabilityTabControl
                         {
                             Margin = new MarginPadding { Top = 120 },
                             Width = 400,
@@ -80,6 +80,7 @@ namespace osu.Game.Screens.Multi.Screens.Match
                             Margin = new MarginPadding { Top = 200 },
                             Text = @"GAME TYPE",
                             Colour = Color4.White,
+                            Font = @"Exo2.0-Bold",
                         },
                         gameTypeTabs = new GameTypeTabControl
                         {
@@ -103,6 +104,7 @@ namespace osu.Game.Screens.Multi.Screens.Match
                         {
                             Text = @"MAX PARTICIPANTS",
                             Colour = Color4.White,
+                            Font = @"Exo2.0-Bold",
                         },
                         maxParticipantsBox = new SettingsTextBox
                         {
@@ -117,6 +119,7 @@ namespace osu.Game.Screens.Multi.Screens.Match
                             Margin = new MarginPadding { Top = 100 },
                             Text = @"PASSWORD (OPTIONAL)",
                             Colour = Color4.White,
+                            Font = @"Exo2.0-Bold",
                         },
                         passwordBox = new SettingsPasswordTextBox
                         {
@@ -143,7 +146,7 @@ namespace osu.Game.Screens.Multi.Screens.Match
             if (int.TryParse(box.Text, out int number))
                 MaxParticipants.Value = number;
             else
-                box.Text = string.Empty;
+                box.Text = MaxParticipants.Value.ToString();
         }
 
         private void onPasswordCommmit(TextBox box, bool newText) => Password.Value = box.Text;
@@ -155,216 +158,6 @@ namespace osu.Game.Screens.Multi.Screens.Match
                 base.OnFocusLost(state);
 
                 OnCommit?.Invoke(this, true);
-            }
-        }
-
-        private class SettingsPasswordTextBox : OsuPasswordTextBox
-        {
-            private readonly Box labelBox;
-
-            public SettingsPasswordTextBox()
-            {
-                TextContainer.Margin = new MarginPadding { Left = 80 };
-
-                Content.Add(new Container
-                {
-                    Masking = true,
-                    CornerRadius = 3,
-                    RelativeSizeAxes = Axes.Y,
-                    Width = 80,
-                    Children = new Drawable[]
-                    {
-                        labelBox = new Box
-                        {
-                            RelativeSizeAxes = Axes.Both,
-                            Colour = Color4.Black,
-                        },
-                        new OsuSpriteText
-                        {
-                            Text = @"Password",
-                            Anchor = Anchor.Centre,
-                            Origin = Anchor.Centre,
-                        },
-                    }
-                });
-            }
-
-            [BackgroundDependencyLoader]
-            private void load(OsuColour colours)
-            {
-                labelBox.Colour = colours.Gray4;
-            }
-
-            protected override void OnFocusLost(InputState state)
-            {
-                base.OnFocusLost(state);
-
-                OnCommit?.Invoke(this, true);
-            }
-        }
-
-        private class AvailabilityTabControl : OsuTabControl<RoomAvailability>
-        {
-            protected override TabItem<RoomAvailability> CreateTabItem(RoomAvailability value) => new AvailabilityTabItem(value);
-
-            private class AvailabilityTabItem : OsuTabItem
-            {
-                private readonly Box bg;
-
-                private OsuColour colours;
-
-                public AvailabilityTabItem(RoomAvailability value)
-                    : base(value)
-                {
-                    AutoSizeAxes = Axes.None;
-                    CornerRadius = 7;
-                    Masking = true;
-                    Width = 110;
-
-                    Children = new Drawable[]
-                    {
-                        bg = new Box
-                        {
-                            Colour = Color4.Black,
-                            RelativeSizeAxes = Axes.Both,
-                        },
-                        new OsuSpriteText
-                        {
-                            Text = value.GetDescription(),
-                            Colour = Color4.White,
-                            Anchor = Anchor.Centre,
-                            Origin = Anchor.Centre,
-                        }
-                    };
-                }
-
-                [BackgroundDependencyLoader]
-                private void load(OsuColour colours)
-                {
-                    this.colours = colours;
-
-                    bg.Colour = Active.Value ? colours.Green : colours.Gray4;
-                }
-                protected override bool OnHover(InputState state)
-                {
-                    bg.FadeTo(0.8f, 200);
-
-                    return base.OnHover(state);
-                }
-
-                protected override void OnHoverLost(InputState state)
-                {
-                    bg.FadeTo(1f, 200);
-
-                    base.OnHoverLost(state);
-                }
-
-                private void fadeActive()
-                {
-                    bg.FadeColour(colours.Green, 150, Easing.InSine);
-                }
-
-                private void fadeInactive()
-                {
-                    bg.FadeColour(colours.Gray4, 150, Easing.InSine);
-                }
-
-                protected override void OnActivated()
-                {
-                    fadeActive();
-
-                    base.OnActivated();
-                }
-
-                protected override void OnDeactivated()
-                {
-                    fadeInactive();
-
-                    base.OnDeactivated();
-                }
-            }
-        }
-
-        private class GameTypeTabControl : OsuTabControl<GameType>
-        {
-            protected override TabItem<GameType> CreateTabItem(GameType value) => new GameTypeTabItem(value);
-
-            protected override Dropdown<GameType> CreateDropdown() => null;
-
-            private readonly OsuSpriteText activeTabText;
-
-            public GameTypeTabControl()
-            {
-                AddInternal(activeTabText = new OsuSpriteText
-                {
-                    Anchor = Anchor.BottomLeft,
-                    Origin = Anchor.BottomLeft,
-                });
-            }
-
-            [BackgroundDependencyLoader]
-            private void load(OsuColour colours)
-            {
-                activeTabText.Colour = colours.Yellow;
-            }
-
-            protected override void SelectTab(TabItem<GameType> tab)
-            {
-                base.SelectTab(tab);
-
-                activeTabText.Text = tab.Value.GetDescription();
-            }
-
-            private class GameTypeTabItem : OsuTabItem
-            {
-                private readonly DrawableGameType drawableType;
-
-                public GameTypeTabItem(GameType value)
-                    : base(value)
-                {
-                    AutoSizeAxes = Axes.None;
-                    RelativeSizeAxes = Axes.None;
-
-                    Child = drawableType = new DrawableGameType(value, 24)
-                    {
-                        RelativeSizeAxes = Axes.Both,
-                    };
-
-                    Height = 60;
-                    Width = 60;
-
-                    Colour = Color4.White;
-                }
-
-                [BackgroundDependencyLoader]
-                private void load(OsuColour colours)
-                {
-                    drawableType.BorderColour = colours.Yellow;
-                }
-
-                private void fadeActive()
-                {
-                    drawableType.BorderThickness = 5;
-                }
-
-                private void fadeInactive()
-                {
-                    drawableType.BorderThickness = 0;
-                }
-
-                protected override void OnActivated()
-                {
-                    base.OnActivated();
-
-                    fadeActive();
-                }
-
-                protected override void OnDeactivated()
-                {
-                    base.OnDeactivated();
-
-                    fadeInactive();
-                }
             }
         }
     }
